web_app/requirements.txt
rdflib
Werkzeug==2.2.2
Flask<=2.2.0
SPARQLWrapper
lxml
nltk>=3.4
pytest>=3.0.7
PyYAML>=3.13
regex>=2018
requests>=2.14.2
spacy
flask-jwt-extended
flasgger
psycopg2-binary
validators
flask_cors
google-cloud-storage
google-cloud-dataproc
Flask-Limiter
sendgrid
scikit-learn
elasticsearch
gunicorn
